backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
colorama
google-generativeai
pydantic